requirements.txt
anthropic>=0.40
requests>=2.31
python-dotenv>=1.0
gspread>=6.0
google-auth>=2.30
pydantic>=2.0
